% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\docType{data}
\name{csv_io}
\alias{csv_io}
\title{Read and write \code{.csv} files.}
\format{\preformatted{List of 2
 $ read :function (path, args = NULL)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 23 13 26 1 13 1 23 26
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x12fb53150> 
 $ write:function (data, path, args = NULL)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 29 14 33 1 14 1 29 33
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x12fb53150> 
}}
\usage{
csv_io
}
\description{
Read and write \code{.csv} files.
}
\seealso{
\code{\link{datanode}}
}
\keyword{datasets}

