% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datanode.R
\name{datanode}
\alias{datanode}
\title{Cache the result of a given expression.}
\usage{
datanode(path, expr, force = FALSE, depends_on = character(0), ask = F)
}
\arguments{
\item{path}{the file caching the result of \code{expr}}

\item{expr}{the expression to be evaluated if triggered}

\item{force}{whether to force the evaluation of the expression \code{expr}
and the update of its cache, defaults to \code{TRUE}}

\item{depends_on}{a character vector of files on which the evaluation of the
       expression \code{expr} depends on.}
}
\value{
the result of the evaluation of the expression \code{expr} if
        triggered, or its cached value stored in \code{path} otherwise
}
\description{
Evaluates an expression and caches its result. If a cache exists, the
expression is re-evaluated only if one of the dependency files is more
recent than the cache.
}
\details{
The given expression \code{expr} is evaluated only if one of the
following condition is true:

\itemize {
 - the file identified in \code{path} does *not* exist; or
 - \code{force} is \code{TRUE}; or
 - the latest modified time of the files in \code{depends_on} is later than
   the modified time of the file in \code{path}; or
 - \code{ask} is \code{TRUE} and the user answers \code{y} at the console
   (only considered in interactive mode).
}
}

