% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\docType{data}
\name{rdata_io}
\alias{rdata_io}
\title{Read and write \code{.RData} files.}
\format{\preformatted{List of 2
 $ read :function (path, args = NULL)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 1 15 6 1 15 1 1 6
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x12fb53150> 
 $ write:function (data, path, args = NULL)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 8 16 12 1 16 1 8 12
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x12fb53150> 
}}
\usage{
rdata_io
}
\description{
Read and write \code{.RData} files.
}
\seealso{
\code{\link{datanode}}
}
\keyword{datasets}

