% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\docType{data}
\name{rds_io}
\alias{rds_io}
\title{Read and write \code{.rds} files.}
\format{\preformatted{List of 2
 $ read :function (path, args = NULL)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 43 13 46 1 13 1 43 46
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x12fb53150> 
 $ write:function (data, path, args = NULL)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 48 14 52 1 14 1 48 52
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x12fb53150> 
}}
\usage{
rds_io
}
\description{
Read and write \code{.rds} files.
}
\seealso{
\code{\link{datanode}}
}
\keyword{datasets}

