% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datanode.R
\name{datanode}
\alias{datanode}
\title{Cache the result of a given expression.}
\usage{
datanode(path, expr, force = FALSE, depends_on = character(0), ask = F)
}
\arguments{
\item{path}{the file caching the result of `expr`}

\item{expr}{the expression to be evaluated if triggered}

\item{force}{(default `FALSE`) whether to force the evaluation of the expression `expr` and the update of its cache}

\item{depends_on}{a character vector of files on which the evaluation of the expression `expr` depends on.}
}
\value{
the result of the evaluation of the expression `expr` if triggered, or its cached value stored in `path` otherwise
}
\description{
The given expression `expr` is evaluated only if one of the following condition is true:
}
\details{
- the file identified in `path` does *not* exist; or
- `force` is `TRUE`; or
- the latest modified time of the files in `depends_on` is later than the modified time of the file in `path`; or
- `ask` is `TRUE` and the user answers `y` at the console (only considered in interactive mode).
}

